# api/api.yaml
openapi: 3.1.0
info:
    title: Questioner API
    version: "0.1"

servers:
  - url: http://localhost:8080

paths:
    /api/task/{id}:
        get:
            parameters:
              - name: id
                in: path
                required: true
                schema:
                    type: integer
                    format: uint64
            responses:
                '200':
                    description: Task to be worked on
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Task'

    /api/task/{id}/solve:
        get:
            parameters:
              - name: id
                in: path
                required: true
                schema:
                    type: integer
                    format: uint64
              - name: solutionID
                in: query
                required: true
                schema:
                    type: integer
                    format: uint64
            responses:
                '200':
                    description: Check if solution is correct
                    content:
                        application/json:
                            schema:
                                type: boolean

    /api/task/{id}/resolve:
        get:
            parameters:
              - name: id
                in: path
                required: true
                schema:
                    type: integer
                    format: uint64
            responses:
                '200':
                    description: Correct answers
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Answer'

components:
    schemas:
        Task:
            type: object
            properties:
                id:
                    type: integer
                    format: uint64
                question:
                    type: string
                answers:
                    type: array
                    items:
                        $ref: '#/components/schemas/Answer'

        Answer:
            type: object
            properties:
                id:
                    type: integer
                    format: uint64
                text:
                    type: string